Crate vergen

Defines the vergen function.

vergen when used in conjunction with the build script support from cargo, generates a file in OUT_DIR (defined by cargo) with three functions defined (now, sha, and semver). This file can then be use with include! to pull the functions into your source for use.

Example Cargo.toml

[package]
build = "build.rs"

[build-dependencies]
vergen = "*"

Example build.rs

// build.rs
extern crate vergen;

use vergen::vergen;

fn main() {
    vergen();
}

Example Usage

// Other stuff
include!(concat!(env!("OUT_DIR"), "/version.rs"));

// Example version function
fn version() -> String {
   format!("{} {} blah {}", now(), sha(), semver())
}

Structs

Flags

Constants

COMMIT_DATE

NOW

SEMVER

SHA

SHORT_NOW

SHORT_SHA

TARGET

Functions

vergen
Create the version.rs file in OUT_DIR, and write three functions into it.